requirements.txt
openai==0.27.7
python-dotenv==1.0.0
langchain==0.0.178
google-search-results==2.4.2
tiktoken==0.4.0